moved the new radio buttons to a different file
